requirements_pix2sgg.txt
accelerate==0.22.0
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
cfgv==3.4.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpickle==2.2.1
comm==0.1.4
contexttimer==0.3.3
contourpy==1.1.0
cryptography @ file:///croot/cryptography_1689373676338/work
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
distlib==0.3.7
easydict==1.10
einops==0.6.1
exceptiongroup==1.1.3
executing==1.2.0
fairscale==0.4.4
fastjsonschema==2.18.0
filelock==3.12.2
fonttools==4.42.1
fqdn==1.5.1
fsspec==2023.6.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.32
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
h5py==3.9.0
huggingface-hub==0.16.4
identify==2.5.27
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.1
imantics==0.1.12
importlib-metadata==6.8.0
importlib-resources==6.0.1
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.25.1
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kaggle==1.5.16
kiwisolver==1.4.4
lazy_loader==0.3
lxml==4.9.3
markdown-it-py==3.0.0
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mkl-fft==1.3.6
mkl-service==2.4.0
msgspec==0.18.1
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
nodeenv==1.8.0
notebook==7.0.2
notebook_shim==0.2.3
omegaconf==2.3.0
opencv-python==4.8.0.76
opendatasets==0.1.22
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
peft==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.16.1
portalocker==2.7.0
pre-commit==3.3.3
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
regex==2023.8.8
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
streamlit==1.25.0
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
-e git+https://github.com/huggingface/transformers.git@ba7054533fa455e8b2dd35feb077e0c7aae646b3#egg=transformers
triton==2.0.0
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.3.0
validators==0.21.2
virtualenv==20.24.3
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.6.2
widgetsnbextension==4.0.8
xmljson==0.2.1
zipp==3.16.2
